requirements.txt
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.2
langgraph>=0.0.20
chromadb>=0.4.22
tiktoken>=0.5.2
beautifulsoup4>=4.12.2
langchain-text-splitters>=0.0.1
pydantic>=2.0.3
typing_extensions>=4.8.0
langchainhub>=0.1.13
python-docx>=0.8.11
langchain-deepseek>=0.0.1 